Renaming marklogic-junit to marklogic-junit5
This allows for marklogic-junit5 to be published under marklogic-community
