+++ marklogic-junit5/examples/simple-dhf4/build.gradle
@@ -10,9 +10,9 @@ repositories {
 }
 
 dependencies {
-  mlRestApi "com.marklogic:marklogic-unit-test-modules:0.13.develop"
+  mlRestApi "com.marklogic:marklogic-unit-test-modules:1.0.beta"
 
-	testCompile "com.marklogic:marklogic-junit:0.13.develop"
+	testCompile "com.marklogic:marklogic-junit5:1.0.beta"
   testCompile "com.marklogic:marklogic-data-hub:4.2.2"
 
 	// Needed by Gradle 4.6+
