+++ marklogic-junit5/examples/simple-dhf4/gradle/wrapper/gradle-wrapper.properties
