+++ marklogic-junit5/examples/simple-dhf4/src/test/ml-modules/root/test/suites/My Tests/simple-equals-test.xqy
