+++ marklogic-junit5/examples/simple-ml-gradle/.gitignore
