+++ marklogic-junit5/examples/simple-ml-gradle/build.gradle
@@ -15,11 +15,11 @@ repositories {
 }
 
 dependencies {
-  mlBundle "com.marklogic:marklogic-unit-test-modules:0.13.develop"
+  mlBundle "com.marklogic:marklogic-unit-test-modules:1.0.beta"
 
 	compile "com.marklogic:marklogic-client-api:4.2.0"
 
-	testCompile "com.marklogic:marklogic-junit:0.13.develop"
+	testCompile "com.marklogic:marklogic-junit:1.0.beta"
 
 	// Needed by Gradle 4.6+ - see https://www.petrikainulainen.net/programming/testing/junit-5-tutorial-running-unit-tests-with-gradle/
 	testRuntime "org.junit.jupiter:junit-jupiter-engine:5.4.1"
