+++ marklogic-junit5/examples/simple-ml-gradle/gradle/wrapper/gradle-wrapper.properties
