+++ marklogic-junit5/examples/simple-ml-gradle/src/main/java/org/example/EchoManager.java
