+++ marklogic-junit5/examples/simple-ml-gradle/src/main/ml-modules/root/example-lib.xqy
