+++ marklogic-junit5/examples/simple-ml-gradle/src/main/ml-modules/services/echo.sjs
