+++ marklogic-junit5/examples/simple-ml-gradle/src/test/java/org/example/PermissionsTest.java
