+++ marklogic-junit5/examples/simple-ml-gradle/src/test/ml-modules/root/test/suites/My Tests/simple-equals-test.xqy
