+++ marklogic-junit5/examples/simple-ml-gradle/src/test/resources/logback.xml
