+++ marklogic-junit5/gradle.properties
@@ -4,4 +4,3 @@ mlUsername=admin
 mlPassword=changeme-in-gradle-local.properties
 mlHost=localhost
 mlTestRestPort=8000
-
