+++ marklogic-junit5/src/main/java/com/marklogic/junit5/XmlNode.java
