+++ marklogic-junit5/src/test/java/com/marklogic/junit5/spring/XmlNodeTest.java
